update Week 08

+++ Week08-Cmake/src/math/include/fact.h
@@ -1,4 +1,3 @@
 #pragma once
 
-
-int fact(int n);
+long long fact(int n);
